… Cargo.lock
… Cargo.toml
… src/ast/expr.rs
… src/ast/mod.rs
… src/extract.rs
… src/lib.rs
… src/prelude.rs
… src/scheduler.rs
… src/serialize.rs
… src/sort/add_primitive/src/lib.rs
… src/sort/bigint.rs
… src/sort/bigrat.rs
… src/sort/f64.rs
… src/sort/fn.rs
… src/sort/i64.rs
… src/sort/map.rs
… src/sort/mod.rs
… src/sort/multiset.rs
… src/sort/set.rs
… src/sort/string.rs
… src/sort/vec.rs
… src/termdag.rs
… src/typechecking.rs
… src/util.rs
… tests/integration_test.rs