This PR removes our dependency on `Symbol`. It is unnecessary because:
- the new backend does its own string interning
- I was able to reach performance parity in the frontend