@yihozhang 2025-06-11
I can confirm the speedup compared to main.

On math-microbenchmark
```
main:     time:   [2.0563 s 2.0854 s 2.1165 s]
this PR:     time:   [1.8570 s 1.8834 s 1.9127 s]
```