@ezrosent 2025-06-12
Nice! A few notes on why things might be faster here:

* As noted in the PR description, the new backend does its own interning. That means that the interning from `Symbol` is just adding overhead as we effectively have to unbox twice.
* `Symbol` makes hashing and comparison for strings faster but it makes everything else slower. In particular I recall going from `String` to `Symbol` (note all the "<string>".into() calls we're getting rid of) as you essentially need to do a full hashtable lookup and possible insertion under a lock.